+++ src/lib.rs
@@ -997,7 +997,7 @@ impl<'a, 'input: 'a> Node<'a, 'input> {
 
     /// Returns the previous sibling element of this node.
     pub fn prev_sibling_element(&self) -> Option<Self> {
-        self.prev_siblings().filter(|n| n.is_element()).nth(0)
+        self.prev_siblings().skip(1).filter(|n| n.is_element()).nth(0)
     }
 
     /// Returns the next sibling of this node.
@@ -1008,7 +1008,7 @@ impl<'a, 'input: 'a> Node<'a, 'input> {
 
     /// Returns the next sibling element of this node.
     pub fn next_sibling_element(&self) -> Option<Self> {
-        self.next_siblings().filter(|n| n.is_element()).nth(0)
+        self.next_siblings().skip(1).filter(|n| n.is_element()).nth(0)
     }
 
     /// Returns the first child of this node.
@@ -1051,25 +1051,25 @@ impl<'a, 'input: 'a> Node<'a, 'input> {
         AxisIter { node: Some(*self), next: Node::parent }
     }
 
-    /// Returns an iterator over previous sibling nodes.
+    /// Returns an iterator over previous sibling nodes starting at this node.
     #[inline]
     pub fn prev_siblings(&self) -> AxisIter<'a, 'input> {
         AxisIter { node: Some(*self), next: Node::prev_sibling }
     }
 
-    /// Returns an iterator over next sibling nodes.
+    /// Returns an iterator over next sibling nodes starting at this node.
     #[inline]
     pub fn next_siblings(&self) -> AxisIter<'a, 'input> {
         AxisIter { node: Some(*self), next: Node::next_sibling }
     }
 
-    /// Returns an iterator over first children nodes.
+    /// Returns an iterator over first children nodes starting at this node.
     #[inline]
     pub fn first_children(&self) -> AxisIter<'a, 'input> {
         AxisIter { node: Some(*self), next: Node::first_child }
     }
 
-    /// Returns an iterator over last children nodes.
+    /// Returns an iterator over last children nodes starting at this node.
     #[inline]
     pub fn last_children(&self) -> AxisIter<'a, 'input> {
         AxisIter { node: Some(*self), next: Node::last_child }